* Build.GetBuildPath() no longer cleans the target directory by default. * Turned off AOT for all platforms except IOS where we actually need it.

+++ Scripts/Editor/Build.cs
@@ -16,7 +16,7 @@
 */
 
 #if UNITY_EDITOR
-
+ 
 using UnityEngine;
 using UnityEditor;
 using System.Collections;
@@ -73,7 +73,7 @@ public class Build
     /// </summary>
     /// <param name="target">The BuildTarget.</param>
     /// <returns>The full path to the build.</returns>
-    public static string GetBuildPath( BuildTarget target )
+    public static string GetBuildPath( BuildTarget target, bool bCleanTarget = false )
     {
         string projectName = Path.GetFileNameWithoutExtension( Application.productName );
         if ( target == BuildTarget.StandaloneWindows || target == BuildTarget.StandaloneWindows64 )
@@ -86,7 +86,7 @@ public class Build
             projectName += ".ipa";
 
         string directory = Application.dataPath + "/../Clients/" + target.ToString();
-        if ( Directory.Exists( directory ) )
+        if ( bCleanTarget && Directory.Exists( directory ) )
             Directory.Delete( directory, true );
 
         Directory.CreateDirectory( directory );
@@ -111,13 +111,16 @@ public class Build
     {
         yield return null;
 
-        /// generate the AOT code, wait for it to be compiled..
-        FullSerializer.AotHelpers.BuildAOT();
-        while( EditorApplication.isCompiling )
-            yield return null;
+        if ( BuildTarget == BuildTarget.iOS )
+        {
+            /// generate the AOT code, wait for it to be compiled..
+            FullSerializer.AotHelpers.BuildAOT();
+            while( EditorApplication.isCompiling )
+                yield return null;
+        }
 
         string [] buildScenes = GetBuildScenes();
-        string buildPath = GetBuildPath( BuildTarget );
+        string buildPath = GetBuildPath( BuildTarget, true );
 
         BuildError = string.Empty;
         try {
@@ -128,7 +131,8 @@ public class Build
             BuildError = e.ToString();
         }
 
-        FullSerializer.AotHelpers.CleanAOT();
+        if ( BuildTarget == BuildTarget.iOS )
+            FullSerializer.AotHelpers.CleanAOT();
         IsBuilding = false; 
 
         // if BuildPlayer returned no error, but we can't find the file, flag this build as a failure then..
